+++ src/main/sc/player2023/logic/rating/SearchWindow.java
@@ -4,13 +4,13 @@ package sc.player2023.logic.rating;
  * @author Till Fransson
  * @since 13.11.2022
  */
-public record SearchWindow(double lowerBound, double upperBound) {
+public record SearchWindow(int lowerBound, int upperBound) {
 
     public boolean canBeCutBeta() {
         return lowerBound >= upperBound;
     }
 
-    public boolean canBeCutScore(double score) {
+    public boolean canBeCutScore(int score) {
         return lowerBound < score && score < upperBound;
     }
